// src/participation/enums/participation-error-code.enum.ts
export enum ParticipationErrorCode {
  ALREADY_JOINED = 'ALREADY_JOINED',
  NOT_JOINED = 'NOT_JOINED',
  ALREADY_SUBMITTED = 'ALREADY_SUBMITTED',
  CONTEST_NOT_ACTIVE = 'CONTEST_NOT_ACTIVE',
  ACCESS_DENIED_LEVEL = 'ACCESS_DENIED_LEVEL',
  PARTICIPATION_NOT_FOUND = 'PARTICIPATION_NOT_FOUND',
  NOT_OWNER = 'NOT_OWNER',
  INVALID_ANSWER = 'INVALID_ANSWER',
}